services/data-cleanup/requirements.txt
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
